feat: Separate file with ModularSpawner class
